A function's grants get their own module
The previous commit took `function_controller.py` to 619 lines, past the
600-line ceiling, and the architecture ratchet is right to fail on it. The four
helpers that move are one subject — reading a grantee's grants into a response
and writing an inline `permissions` block — and none of them is a route, so the
routes stay where they were and the controller drops to 491 lines.

They lose their leading underscore in the move. A name that another module
imports is that module's interface whatever it is called, and
`from ... import _apply_function_grants` was saying otherwise.
