+++ lemma-backend/app/modules/function/api/controllers/function_controller.py
@@ -2,15 +2,13 @@
 
 from functools import partial
 from uuid import UUID
-from typing import Any, Awaitable, Callable, Optional
+from typing import Optional
 from fastapi import APIRouter, Depends, Query, Request, status
 
 from app.core.api.dependencies import UoWDep, get_uow_factory
 from app.core.infrastructure.db.uow_factory import UnitOfWorkFactory
 from app.core.authorization.conferral import assert_can_confer
 from app.core.authorization.grants import (
-    apply_inline_workload_grants,
-    list_grantee_resource_grants,
     normalize_pod_resource_grants,
     replace_grantee_resource_grants,
     validate_pod_resource_grant_permissions,
@@ -49,6 +47,11 @@ from app.modules.function.api.dependencies import (
     FunctionResourceEditorDep,
     FunctionResourceViewerDep,
 )
+from app.modules.function.api.controllers.function_grants import (
+    apply_function_grants,
+    function_permissions_response,
+    grants_for_functions,
+)
 from app.modules.workspace.contracts.tooling import (
     invalidate_function_workspace_env_cache,
 )
@@ -63,82 +66,6 @@ router = APIRouter(
 def _to_function_response(function: FunctionEntity) -> FunctionResponse:
     payload = function.model_dump()
     return FunctionResponse.model_validate(payload)
-
-
-async def _apply_function_grants(
-    uow_factory: UnitOfWorkFactory,
-    *,
-    pod_id: UUID,
-    function: FunctionEntity,
-    data,
-    user: UserEntity,
-) -> None:
-    """Apply a create/update payload's inline ``permissions`` block.
-
-    Functions accepted this block on neither verb while agents accepted it on
-    create, so the same bundle payload wired an agent and silently left a
-    function with nothing.
-
-    Its own short unit of work, opened after the code has been applied rather
-    than a request-scoped one held across it: the caller's other work needs no
-    connection, and `use_cases.create_function` already opens its own for the
-    parts that do.
-
-    The invalidation is outside the block, which is a fix and not a tidy-up. In
-    the transaction it both held a pooled connection across a Redis round trip
-    and ran in the wrong order -- a concurrent reader could repopulate the cache
-    from the pre-commit state between the invalidation and the commit, which is
-    the hazard `SqlAlchemyUnitOfWork.after_commit` was written for.
-    """
-
-    assert function.id is not None
-    permissions = getattr(data, "permissions", None)
-    if permissions is None:
-        # An absent block leaves the grants alone, so there is nothing to open a
-        # connection for. `apply_inline_workload_grants` returns False here
-        # without touching the session; checking first means the common create
-        # -- which names no permissions at all -- checks out no connection.
-        return
-    function_id = function.id
-    await _write_then_invalidate(
-        uow_factory,
-        write=lambda session: apply_inline_workload_grants(
-            session,
-            pod_id=pod_id,
-            grantee_type="FUNCTION",
-            grantee_id=function_id,
-            permissions=permissions,
-            created_by_user_id=user.id,
-        ),
-        # The sandbox reads its grants from a cached environment, so skipping
-        # this leaves the function running with the access it had before.
-        invalidate=lambda: invalidate_function_workspace_env_cache(
-            pod_id=pod_id,
-            function_id=function_id,
-        ),
-    )
-
-
-async def _write_then_invalidate(
-    uow_factory: UnitOfWorkFactory,
-    *,
-    write: Callable[[Any], Awaitable[bool]],
-    invalidate: Callable[[], Awaitable[None]],
-) -> None:
-    """Commit ``write``, and only then drop what it made stale.
-
-    Both halves are parameters so the order can be asserted without a database
-    and without a double: the order *is* the behaviour here, and it was wrong.
-    Dropping a cache inside the transaction pins a pooled Postgres connection
-    across a Redis round trip, and races -- a concurrent reader can repopulate
-    the cache from the pre-commit state before the commit lands. See
-    `SqlAlchemyUnitOfWork.after_commit`, which exists for the case where the
-    caller cannot close its transaction early like this one can.
-    """
-    async with uow_factory() as uow:
-        applied = await write(uow.session)
-    if applied:
-        await invalidate()
 
 
 async def _function_action_response(
@@ -195,7 +122,7 @@ async def create_function(
     function = await use_cases.create_function(
         pod_id=pod_id, entity=entity, user_id=user.id, code=data.code, request=request
     )
-    await _apply_function_grants(
+    await apply_function_grants(
         uow_factory, pod_id=pod_id, function=function, data=data, user=user
     )
     return await _function_action_response(function)
@@ -238,7 +165,7 @@ async def list_functions(
         pod_id, user_id, limit, page_token, ctx=ctx
     )
 
-    grants_by_function = await _grants_for_functions(uow, pod_id, functions, include)
+    grants_by_function = await grants_for_functions(uow, pod_id, functions, include)
     return FunctionListResponse(
         items=[
             _function_summary_response(
@@ -252,34 +179,6 @@ async def list_functions(
         limit=limit,
         next_page_token=next_cursor,
     )
-
-
-async def _grants_for_functions(
-    uow: Any,
-    pod_id: UUID,
-    functions: list[Any],
-    include: list[str],
-) -> dict[UUID, list[FunctionResourcePermissionResponse]] | None:
-    """Grants for a whole page of functions, or None when not requested."""
-    if not any(part.strip().lower() == "permissions" for part in include):
-        return None
-    from app.core.authorization.grants import list_grants_for_grantees
-
-    ids = [f.id for f in functions if f.id is not None]
-    grouped = await list_grants_for_grantees(
-        uow.session, pod_id=pod_id, grantee_type="FUNCTION", grantee_ids=ids
-    )
-    return {
-        function_id: [
-            FunctionResourcePermissionResponse(
-                resource_type=resource_type,
-                resource_name=resource_name,
-                permission_ids=sorted(set(permission_ids)),
-            )
-            for (resource_type, resource_name), permission_ids in grants.items()
-        ]
-        for function_id, grants in grouped.items()
-    }
 
 
 @router.get(
@@ -311,7 +210,7 @@ async def get_function(
     response = await _function_action_response(function)
     return FunctionDetailResponse(
         **response.model_dump(),
-        permissions=await _function_permissions_response(
+        permissions=await function_permissions_response(
             uow,
             pod_id=pod_id,
             function=function,
@@ -346,7 +245,7 @@ async def get_function_permissions(
         ctx=ctx,
     )
     assert function is not None
-    return await _function_permissions_response(uow, pod_id=pod_id, function=function)
+    return await function_permissions_response(uow, pod_id=pod_id, function=function)
 
 
 @router.put(
@@ -418,7 +317,7 @@ async def replace_function_permissions(
             function_id=function.id,
         )
     )
-    return await _function_permissions_response(uow, pod_id=pod_id, function=function)
+    return await function_permissions_response(uow, pod_id=pod_id, function=function)
 
 
 @router.patch(
@@ -454,7 +353,7 @@ async def update_function(
         user_id=user.id,
         request=request,
     )
-    await _apply_function_grants(
+    await apply_function_grants(
         uow_factory, pod_id=pod_id, function=function, data=data, user=user
     )
     return await _function_action_response(function)
@@ -561,33 +460,6 @@ async def list_runs(
         items=[FunctionRunSummaryResponse.model_validate(r) for r in runs],
         limit=limit,
         next_page_token=next_cursor,
-    )
-
-
-async def _function_permissions_response(
-    uow: UoWDep,
-    *,
-    pod_id: UUID,
-    function: FunctionEntity,
-) -> FunctionPermissionsResponse:
-    assert function.id is not None
-    grouped = await list_grantee_resource_grants(
-        uow.session,
-        pod_id=pod_id,
-        grantee_type="FUNCTION",
-        grantee_id=function.id,
-    )
-    return FunctionPermissionsResponse(
-        function_id=function.id,
-        function_name=function.name,
-        grants=[
-            FunctionResourcePermissionResponse(
-                resource_type=resource_type,
-                resource_name=resource_name,
-                permission_ids=sorted(set(permission_ids)),
-            )
-            for (resource_type, resource_name), permission_ids in grouped.items()
-        ],
     )
